timelock: adds prom counter metric and server
Adds a prometheus counter metric that counts the number
of errors when dumping the store to disk.  If this ever is >= 1
then alerting should trigger as the disk write failed.  Adds logging
to show the earliest block number to enable a manual restart of the
worker.

Adds a prometheus metrics server on a default port.