fix: add 30s timeout to context used to execute operations onchain
Previously, when executing an operation, the timelock-worker called
`bind.WaitMinted` with a context without an associated timeout. This
meant that when there was a problem with the RPC or the network itself,
the executer goroutine would get stuck waiting indefinitely for the call
to return. And because all operations are called sequentially, the
timelock effectively stopped working.

With the added timeout, the context aborts the network call and the
operation fails. The internal logic of the scheduler will take care of
retrying a few minutes later.